defined cql types in standalone file
