+++ src/types/CQLExecution.d.ts
@@ -1,37 +1,6 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 declare module 'cql-execution' {
-  interface CQLCode {
-    system: string;
-    code: string;
-    display?: string;
-    version?: string;
-  }
-
-  export interface ValueSetMap {
-    [key: string]: {
-      [key: string]: CQLCode[];
-    };
-  }
-
-  export interface Results {
-    patientResults: {
-      [patientId: string]: StatementResults;
-    };
-    unfilteredResults: any;
-    localIdPatientResultsMap: {
-      [patientId: string]: LocalIdResults;
-    };
-  }
-
-  export interface StatementResults {
-    [statementName: string]: any;
-  }
-
-  export interface LocalIdResults {
-    [libraryName: string]: {
-      [localId: string]: any;
-    };
-  }
+  import { Results, ValueSetMap } from './CQLTypes';
 
   export class Library {
     constructor(elm: any, repo?: Repository);
