+++ src/types/CQLTypes.ts
@@ -1,0 +1,32 @@
+export interface CQLCode {
+  system: string;
+  code: string;
+  display?: string;
+  version?: string;
+}
+
+export interface ValueSetMap {
+  [key: string]: {
+    [key: string]: CQLCode[];
+  };
+}
+
+export interface Results {
+  patientResults: {
+    [patientId: string]: StatementResults;
+  };
+  unfilteredResults: any;
+  localIdPatientResultsMap: {
+    [patientId: string]: LocalIdResults;
+  };
+}
+
+export interface StatementResults {
+  [statementName: string]: any;
+}
+
+export interface LocalIdResults {
+  [libraryName: string]: {
+    [localId: string]: any;
+  };
+}
