+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './Calculator';
 export * from './ELMTypes';
 export * from './Enums';
 export * from './ExecutionValueSet';
+export * as CQLTypes from './CQLTypes';
